+++ src/encoder/wrap.cpp
@@ -9,15 +9,15 @@ namespace py = pybind11;
 PYBIND11_MODULE(_encoder, obj)
 {
   py::class_<PoissonEncoder>(obj, "PoissonEncoder")
-    .def(py::init<int, float>())
+    .def(py::init<int, double>())
     .def("encode", &PoissonEncoder::encode);
 
   py::class_<PopSANEncoder>(obj, "PopSANEncoder")
-    .def(py::init<int, int, int, float, float, float, pair<float, float>>())
+    .def(py::init<int, int, int, double, double, double, pair<double, double>>())
     .def("encode", &PopSANEncoder::encode);
 
   py::class_<LIFEncoder>(obj, "LIFEncoder")
-    .def(py::init<int, int, pair<float, float>, float, float, float, float, float, float>())
+    .def(py::init<int, int, pair<double, double>, double, double, double, double, double, double>())
     .def("encode", &LIFEncoder::encode)
     .def("reset", &LIFEncoder::reset);
 }
