+++ src/ext_lsm/synapse.cpp
@@ -21,13 +21,13 @@ DynamicSynapse::DynamicSynapse(const SynapseParams& params, int seed)
   c = params.c;
 
   mt19937 mt(seed);
-  normal_distribution<float> randn_u0(params.u0_mean, params.u0_sd);
-  normal_distribution<float> randn_tau_d(params.tau_d_mean, params.tau_d_sd);
-  normal_distribution<float> randn_tau_f(params.tau_f_mean, params.tau_f_sd);
+  normal_distribution<double> randn_u0(params.u0_mean, params.u0_sd);
+  normal_distribution<double> randn_tau_d(params.tau_d_mean, params.tau_d_sd);
+  normal_distribution<double> randn_tau_f(params.tau_f_mean, params.tau_f_sd);
 
   while (true)
   {
-    float cnd = randn_u0(mt);
+    double cnd = randn_u0(mt);
     if (abs(cnd - params.u0_mean) < params.u0_sd)
     {
       u0 = cnd;
@@ -36,7 +36,7 @@ DynamicSynapse::DynamicSynapse(const SynapseParams& params, int seed)
   }
   while (true)
   {
-    float cnd = randn_tau_d(mt);
+    double cnd = randn_tau_d(mt);
     if (abs(cnd - params.tau_d_mean) < params.tau_d_sd)
     {
       tau_d = cnd;
@@ -45,7 +45,7 @@ DynamicSynapse::DynamicSynapse(const SynapseParams& params, int seed)
   }
   while (true)
   {
-    float cnd = randn_tau_f(mt);
+    double cnd = randn_tau_f(mt);
     if (abs(cnd - params.tau_f_mean) < params.tau_f_sd)
     {
       tau_f = cnd;
@@ -65,9 +65,9 @@ void DynamicSynapse::reset()
   dw = 0.;
 }
 
-void DynamicSynapse::calc_grad(float dt)
+void DynamicSynapse::calc_grad(double dt)
 {
-  float spike = pre_neuron->get_spike();
+  double spike = pre_neuron->get_spike();
   du = (-dt / tau_f) * u + u0 * (1. - u) * spike;
   dx = (dt / tau_d) * (1. - x) - u * x * spike;
   dw = (pre_neuron->get_trace() * c * exp(-w) - 1.) * dt;
@@ -83,7 +83,7 @@ void DynamicSynapse::update()
   // cout << "u: " << u << ", x: " << x << ", w: " << w << ", amp: " << amp << endl;
 }
 
-float DynamicSynapse::get_amp()
+double DynamicSynapse::get_amp()
 {
   return amp;
 }
