+++ src/ext_lsm/synapse.hpp
@@ -7,9 +7,9 @@ class BaseNeuron;
 struct SynapseParams
 {
 public:
-  float u0_mean, tau_d_mean, tau_f_mean;
-  float u0_sd, tau_d_sd, tau_f_sd;
-  float c;
+  double u0_mean, tau_d_mean, tau_f_mean;
+  double u0_sd, tau_d_sd, tau_f_sd;
+  double c;
 
   SynapseParams();
 };
@@ -17,11 +17,11 @@ public:
 class DynamicSynapse
 {
   // const
-  float u0, tau_d, tau_f, c;
+  double u0, tau_d, tau_f, c;
 
   // mutable
-  float u, x, w, amp;
-  float du, dx, dw;
+  double u, x, w, amp;
+  double du, dx, dw;
 
 public:
   BaseNeuron* pre_neuron;
@@ -31,9 +31,9 @@ public:
 
   void reset();
 
-  void calc_grad(float dt);
+  void calc_grad(double dt);
 
   void update();
 
-  float get_amp();
+  double get_amp();
 };
